readding player pos and data game object. newly adding player game object though it's currently not being used
the player game object is not fully done, so it doesn't work and just gives a continuous error as the game runs. It should be fixed with a new script made specifically for the player. The save load only runs the default, I'm thinking I'm gonna have to change it from save to a file to another save/load mechanic. As it is now, it semi runs and doesn't cause issues so I'll leave it alone for now.